third_party/blink/web_tests/external/wpt/css/css-properties-values-api/at-property-animation-expected.txt

This is a testharness.js-based test.
PASS @keyframes works with @property
PASS @keyframes picks up the latest @property in the document
PASS Ongoing animation picks up redeclared custom property
PASS Ongoing animation matches new keyframes against the current registration
PASS Ongoing animation picks up redeclared intial value
PASS Ongoing animation picks up redeclared inherits flag
PASS Ongoing animation picks up redeclared meaning of 'unset'
PASS Transitioning from initial value
PASS Transitioning from specified value
PASS Transition triggered by initial value change
PASS No transition when changing types
FAIL No transition when removing @property rule assert_equals: expected "100px" but got " 100px"
FAIL Unregistered properties referencing animated properties update correctly. assert_equals: expected "150px" but got " 150px"
FAIL Registered properties referencing animated properties update correctly. assert_equals: expected "150px" but got " 150px"
Harness: the test ran to completion.
